refactor: api endpoint 수정 반영
